Alter the code to work in terms of codepoints, not characters.
This adds a dependency on an external library, which handles the transformation
from Lisp strings to code point vectors. It likely causes non-trivial slowdown,
but that hasn't been measured at this point.

These changes result in a 100% pass rate for the tests in ABCL. It likely does
the same for Allegro, but has not been tested.